Architectural updates to the api tests and somehow a _tmp directory got in the git index.
